Remove custom item UUID
This allows items of the same type created at different times to stack.

The continuous ability item uses are now also tied to the user's UUID instead of the item's.

The whole idea of unique ids on a per item basis doesn't work very well with the item stack system and would need to be reinvented if needed at some point.