Fix config reload
Config didn't reload correctly and didn't read new values from the yml files if changed. This happened because the config fields in GimmeTag.java were read into Config.java statically at the start and didn't react to the fields being set to new objects.

The fix is to, instead of loading in the config objects directly, create functions that return whatever is the current values in the fields of GimmeTag.java.